INFIVERSE-HR-PLATFORM/backend/runtime-core/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pyjwt>=2.8.0
bcrypt>=4.0.1
python-multipart>=0.0.6
pyotp>=2.9.0
qrcode>=7.4.2
pillow>=10.0.0
cryptography>=41.0.7
psutil>=5.9.5
httpx>=0.25.0
pymongo>=4.6.0
motor>=3.3.2
redis>=5.0.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dateutil>=2.8.2
aiofiles>=23.2.1